Renamed internal.tsx to _internal.tsx
Just to add further clarification that stuff in this file won't actually be available to end-users. Or if they do access it, expect things to break between versions.
